move fidelity to qis
