+++ cirq/__init__.py
@@ -120,7 +120,6 @@ from cirq.linalg import (
     diagonalize_real_symmetric_matrix,
     dot,
     expand_matrix_in_orthogonal_basis,
-    fidelity,
     hilbert_schmidt_inner_product,
     is_diagonal,
     is_hermitian,
@@ -302,6 +301,7 @@ from cirq.qis import (
     density_matrix_from_state_vector,
     dirac_notation,
     eye_tensor,
+    fidelity,
     one_hot,
     STATE_VECTOR_LIKE,
     to_valid_density_matrix,
